use double for sentimentScore

+++ portfolio/src/main/java/com/google/sps/data/Message.java
@@ -4,9 +4,9 @@ public final class Message {
   private final String name;
   private final String email;
   private final String message;
-  private final float sentimentScore;
+  private final Double sentimentScore;
 
-  public Message(String name, String email, String message, float sentimentScore) {
+  public Message(String name, String email, String message, Double sentimentScore) {
     this.name = name;
     this.email = email;
     this.message = message;
@@ -25,7 +25,7 @@ public final class Message {
     return this.message;
   }
 
-  public float getSentimentScore() {
+  public Double getSentimentScore() {
     return this.sentimentScore;
   }
 }
